requirements.txt
pandora-chatgpt
redis